// TyrNext/app/store/actions/interfaces.ts
export const INIT_STATE = 'INIT_STATE';
export const RESET_STATE = 'RESET_STATE';
export const STARTUP_WITH_AUTH = 'STARTUP_WITH_AUTH';
export const START_NEW_GAME = 'START_NEW_GAME';
// ...
export const ADD_YAKU = 'ADD_YAKU';
export const INIT_REQUIRED_YAKU = 'INIT_REQUIRED_YAKU';
export const REMOVE_YAKU = 'REMOVE_YAKU';
// ...
export const TOGGLE_RIICHI_NOTIFICATION = 'TOGGLE_RIICHI_NOTIFICATION';

interface InitStateAction {
  type: typeof INIT_STATE;
}

interface ResetStateAction {
  type: typeof RESET_STATE;
}

interface StartupWithAuthAction {
  type: typeof STARTUP_WITH_AUTH;
  payload: {
    token: string;
    personId: number;
    sessionId: string; // analytics session id, not gaming session!
  };
}

interface StartNewGameAction {
  type: typeof START_NEW_GAME;
  payload?: number[];
}

interface GoToCurrentGameAction {
  type: typeof GO_TO_CURRENT_GAME;
}

interface GoToDonateAction {
  type: typeof GO_TO_DONATE;
}

interface GoToEventSelectAction {
  type: typeof GOTO_EVENT_SELECT;
}

interface SearchPlayerAction {
  type: typeof SEARCH_PLAYER;
  payload: IAppState['newGameSelectedPlayerSide'];
}

interface ShowLastResultsAction {
  type: typeof SHOW_LAST_RESULTS;
}

interface ShowGameLogAction {
  type: typeof SHOW_GAME_LOG;
}

interface OpenSettingsAction {
  type: typeof OPEN_SETTINGS;
}

interface GotoNextScreenAction {
  type: typeof GOTO_NEXT_SCREEN;
}

interface GotoPrevScreenAction {
  type: typeof GOTO_PREV_SCREEN;
}

interface SetDoraCountAction {
  type: typeof SET_DORA_COUNT;
  payload: {
    count: number;
    winner?: number;
  };
}

interface SetFuCountAction {
  type: typeof SET_FU_COUNT;
  payload: {
    count: number;
    winner?: number;
  };
}

interface AddYakuAction {
  type: typeof ADD_YAKU;
  payload: {
    id: YakuId;
    winner?: number;
  };
}

interface InitRequiredYakuAction {
  type: typeof INIT_REQUIRED_YAKU;
}

interface RemoveYakuAction {
  type: typeof REMOVE_YAKU;
  payload: {
    id: YakuId;
    winner?: number;
  };
}

interface ToggleRiichiAction {
  type: typeof TOGGLE_RIICHI;
  payload: number;
}

interface ToggleWinnerAction {
  type: typeof TOGGLE_WINNER;
  payload: number;
}

interface ToggleLoserAction {
  type: typeof TOGGLE_LOSER;
  payload: number;
}

interface SelectPaoAction {
  type: typeof SELECT_PAO;
  payload: {
    paoId: number;
    winnerId: number;
  };
}

interface ToggleDeadhandAction {
  type: typeof TOGGLE_DEADHAND;
  payload: number;
}

interface ToggleNagashiAction {
  type: typeof TOGGLE_NAGASHI;
  payload: number;
}

interface LoginActionInit {
  type: typeof LOGIN_INIT;
  payload: {
    email: string;
    password: string;
    sessionId: string;
  };
}
interface LoginActionSuccess {
  type: typeof LOGIN_SUCCESS;
  payload: AuthAuthorizeResponse;
}
interface LoginActionFail {
  type: typeof LOGIN_FAIL;
  payload: RemoteError;
}
interface ResetLoginErrorAction {
  type: typeof RESET_LOGIN_ERROR;
}
interface SetCredentialsAction {
  type: typeof SET_CREDENTIALS;
  payload: {
    authToken: string;
    personId: number;
    sessionId: string;
  };
}
interface UpdateCurrentGamesActionInit {
  type: typeof UPDATE_CURRENT_GAMES_INIT;
}
interface UpdateCurrentGamesActionSuccess {
  type: typeof UPDATE_CURRENT_GAMES_SUCCESS;
  payload: {
    games: CurrentSession[];
    gameConfig: GameConfig;
    timerState: EventsGetTimerStateResponse;
  };
}
interface UpdateCurrentGamesActionFail {
  type: typeof UPDATE_CURRENT_GAMES_FAIL;
  payload: RemoteError;
}
interface GetGameOverviewActionInit {
  type: typeof GET_GAME_OVERVIEW_INIT;
  payload: string;
}
interface GetGameOverviewActionSuccess {
  type: typeof GET_GAME_OVERVIEW_SUCCESS;
  payload: GamesGetSessionOverviewResponse;
}
interface GetGameOverviewActionFail {
  type: typeof GET_GAME_OVERVIEW_FAIL;
  payload: RemoteError;
}
interface GetUserinfoActionInit {
  type: typeof GET_USERINFO_INIT;
  payload: number;
}
interface GetUserinfoActionSuccess {
  type: typeof GET_USERINFO_SUCCESS;
  payload: PersonEx;
}
interface GetUserinfoActionFail {
  type: typeof GET_USERINFO_FAIL;
  payload: RemoteError;
}
export interface ForceLogoutAction {
  type: typeof FORCE_LOGOUT;
  payload: RemoteError | undefined;
}
interface GetOtherTablesListActionInit {
  type: typeof GET_OTHER_TABLES_LIST_INIT;
}
interface GetOtherTablesListActionReload {
  type: typeof GET_OTHER_TABLES_LIST_RELOAD;
}
interface GetOtherTablesListActionSuccess {
  type: typeof GET_OTHER_TABLES_LIST_SUCCESS;
  payload: TableState[];
}
interface GetOtherTablesListActionFail {
  type: typeof GET_OTHER_TABLES_LIST_FAIL;
  payload: RemoteError;
}
interface GetOtherTableActionInit {
  type: typeof GET_OTHER_TABLE_INIT;
  payload: string;
}
interface GetOtherTableActionReload {
  type: typeof GET_OTHER_TABLE_RELOAD;
}
interface GetOtherTableActionSuccess {
  type: typeof GET_OTHER_TABLE_SUCCESS;
  payload: GamesGetSessionOverviewResponse;
}
interface GetOtherTableActionFail {
  type: typeof GET_OTHER_TABLE_FAIL;
  payload: RemoteError;
}
interface GetAllRoundsActionInit {
  type: typeof GET_ALL_ROUNDS_INIT;
  payload: string;
}
interface GetAllRoundsActionSuccess {
  type: typeof GET_ALL_ROUNDS_SUCCESS;
  payload: RoundState[];
}
interface GetAllRoundsActionFail {
  type: typeof GET_ALL_ROUNDS_FAIL;
  payload: RemoteError;
}
interface SetTimerAction {
  type: typeof SET_TIMER;
  payload: {
    waiting: boolean;
    secondsRemaining: number;
    autostartSecondsRemaining: number;
    haveAutostart: boolean;
  };
}
interface UpdateTimerDataAction {
  type: typeof UPDATE_TIMER_DATA;
  payload: {
    waiting: boolean;
    secondsRemaining?: number;
    lastUpdateTimestamp?: number;
    autostartSecondsRemaining?: number;
    autostartLastUpdateTimestamp?: number;
  };
}

interface GetChangesOverviewActionInit {
  type: typeof GET_CHANGES_OVERVIEW_INIT;
  payload: IAppState;
}

interface GetChangesOverviewActionSuccess {
  type: typeof GET_CHANGES_OVERVIEW_SUCCESS;
  payload: RoundState;
}

interface GetChangesOverviewActionFail {
  type: typeof GET_CHANGES_OVERVIEW_FAIL;
  payload: RemoteError;
}

interface GetLastResultsActionInit {
  type: typeof GET_LAST_RESULTS_INIT;
}

interface GetLastResultsActionSuccess {
  type: typeof GET_LAST_RESULTS_SUCCESS;
  payload: SessionHistoryResult[];
}

interface GetLastResultsActionFail {
  type: typeof GET_LAST_RESULTS_FAIL;
  payload: RemoteError;
}

interface GetAllPlayersActionInit {
  type: typeof GET_ALL_PLAYERS_INIT;
}

interface GetAllPlayersActionSuccess {
  type: typeof GET_ALL_PLAYERS_SUCCESS;
  payload: RegisteredPlayer[];
}

interface GetAllPlayersActionFail {
  type: typeof GET_ALL_PLAYERS_FAIL;
  payload: RemoteError;
}

interface StartGameActionInit {
  type: typeof START_GAME_INIT;
  payload: number[];
}

interface StartGameActionSuccess {
  type: typeof START_GAME_SUCCESS;
  payload: string;
}

interface StartGameActionFail {
  type: typeof START_GAME_FAIL;
  payload: RemoteError;
}

interface AddRoundActionInit {
  type: typeof ADD_ROUND_INIT;
  payload: IAppState;
}

interface AddRoundActionSuccess {
  type: typeof ADD_ROUND_SUCCESS;
  payload: GamesAddRoundResponse;
}

interface AddRoundActionFail {
  type: typeof ADD_ROUND_FAIL;
  payload: RemoteError;
}

interface EventsGetListActionInit {
  type: typeof EVENTS_GET_LIST_INIT;
}

export interface EventsGetListActionSuccess {
  type: typeof EVENTS_GET_LIST_SUCCESS;
  payload: MyEvent[];
}

interface EventsGetListActionFail {
  type: typeof EVENTS_GET_LIST_FAIL;
  payload: RemoteError;
}

interface SelectEventAction {
  type: typeof SELECT_EVENT;
  payload: number;
}

interface InitBlankOutcomeAction {
  type: typeof INIT_BLANK_OUTCOME;
  payload: RoundOutcome;
}

interface RandomizeNewgamePlayersAction {
  type: typeof RANDOMIZE_NEWGAME_PLAYERS;
}

interface ClearNewgamePlayersAction {
  type: typeof CLEAR_NEWGAME_PLAYERS;
}

interface SetNewgamePlayersAction {
  type: typeof SET_NEWGAME_PLAYERS;
}

interface SelectNewgameEastAction {
  type: typeof SELECT_NEWGAME_PLAYER_EAST;
  payload: number;
}

interface SelectNewgameSouthAction {
  type: typeof SELECT_NEWGAME_PLAYER_SOUTH;
  payload: number;
}

interface SelectNewgameWestAction {
  type: typeof SELECT_NEWGAME_PLAYER_WEST;
  payload: number;
}

interface SelectNewgameNorthAction {
  type: typeof SELECT_NEWGAME_PLAYER_NORTH;
  payload: number;
}

interface ToggleOverviewDiffbyAction {
  type: typeof TOGGLE_OVERVIEW_DIFFBY;
  payload: number;
}

interface ToggleAdditionalTableInfoAction {
  type: typeof TOGGLE_ADDITIONAL_TABLE_INFO;
}

interface TableRotateClockwiseAction {
  type: typeof TABLE_ROTATE_CLOCKWISE;
}

interface TableRotateCounterclockwiseAction {
  type: typeof TABLE_ROTATE_COUNTERCLOCKWISE;
}

interface SettingsSaveThemeAction {
  type: typeof SETTINGS_SAVE_THEME;
  payload: string;
}

interface SettingsSaveLangAction {
  type: typeof SETTINGS_SAVE_LANG;
  payload: string;
}

interface SettingsSaveSingleDeviceMode {
  type: typeof SETTINGS_SAVE_SINGLE_DEVICE_MODE;
  payload: boolean;
}

interface UpdateStateSettingsAction {
  type: typeof UPDATE_STATE_SETTINGS;
}

interface SetStateSettingsAction {
  type: typeof SET_STATE_SETTINGS;
  payload: { [key: string]: any } | undefined;
}

interface TrackArbitraryEventAction {
  type: typeof TRACK_ARBITRARY_EVENT;
  payload: [string, { [key: string]: any }];
}

interface TrackScreenEnterAction {
  type: typeof TRACK_SCREEN_ENTER;
  payload: string;
}

interface HistoryInitAction {
  type: typeof HISTORY_INIT;
}

interface CallRefereeAction {
  type: typeof CALL_REFEREE;
}

interface ToggleRiichiNotificationAction {
  type: typeof TOGGLE_RIICHI_NOTIFICATION;
  payload: boolean;
}

export type AppActionTypes =
  | InitStateAction
  | ResetStateAction
  | StartupWithAuthAction
  | StartNewGameAction
  | GoToCurrentGameAction
  | GoToEventSelectAction
  | GoToDonateAction
  | SearchPlayerAction
  | ShowLastResultsAction
  | ShowGameLogAction
  | OpenSettingsAction
  | GotoNextScreenAction
  | GotoPrevScreenAction
  | SetDoraCountAction
  | SetFuCountAction
  | AddYakuAction
  | RemoveYakuAction
  | InitRequiredYakuAction
  | ToggleRiichiAction
  | ToggleWinnerAction
  | ToggleLoserAction
  | SelectPaoAction
  | ToggleDeadhandAction
  | ToggleNagashiAction
  | LoginActionSuccess
  | ResetLoginErrorAction
  | UpdateCurrentGamesActionSuccess
  | GetGameOverviewActionSuccess
  | LoginActionFail
  | UpdateCurrentGamesActionFail
  | GetGameOverviewActionFail
  | SetCredentialsAction
  | ForceLogoutAction
  | GetOtherTableActionSuccess
  | GetOtherTableActionFail
  | GetOtherTablesListActionSuccess
  | GetOtherTablesListActionFail
  | GetAllRoundsActionInit
  | GetAllRoundsActionSuccess
  | GetAllRoundsActionFail
  | GetUserinfoActionInit
  | GetUserinfoActionSuccess
  | GetUserinfoActionFail
  | SetTimerAction
  | UpdateTimerDataAction
  | GetChangesOverviewActionSuccess
  | GetChangesOverviewActionFail
  | GetLastResultsActionSuccess
  | GetLastResultsActionFail
  | GetAllPlayersActionSuccess
  | GetAllPlayersActionFail
  | StartGameActionSuccess
  | StartGameActionFail
  | LoginActionInit
  | UpdateCurrentGamesActionInit
  | GetGameOverviewActionInit
  | GetOtherTableActionInit
  | GetOtherTableActionReload
  | GetOtherTablesListActionInit
  | GetOtherTablesListActionReload
  | GetChangesOverviewActionInit
  | GetLastResultsActionInit
  | GetAllPlayersActionInit
  | StartGameActionInit
  | AddRoundActionInit
  | AddRoundActionSuccess
  | AddRoundActionFail
  | InitBlankOutcomeAction
  | RandomizeNewgamePlayersAction
  | ClearNewgamePlayersAction
  | SelectNewgameNorthAction
  | SetNewgamePlayersAction
  | SelectNewgameEastAction
  | SelectNewgameSouthAction
  | SelectNewgameWestAction
  | ToggleOverviewDiffbyAction
  | ToggleAdditionalTableInfoAction
  | TableRotateClockwiseAction
  | TableRotateCounterclockwiseAction
  | SettingsSaveLangAction
  | SettingsSaveSingleDeviceMode
  | SettingsSaveThemeAction
  | UpdateStateSettingsAction
  | SetStateSettingsAction
  | TrackArbitraryEventAction
  | TrackScreenEnterAction
  | EventsGetListActionInit
  | EventsGetListActionSuccess
  | EventsGetListActionFail
  | SelectEventAction
  | HistoryInitAction
  | CallRefereeAction
  | ToggleRiichiNotificationAction;